Added pages/camera to branch
